+++ include/mqtt_client/MqttClient_ros2.h
@@ -33,19 +33,8 @@ SOFTWARE.
 #include <string>
 
 #include <mqtt/async_client.h>
-
-#include "mqtt_client/srv/is_connected.hpp"
-#include "rclcpp/logger.hpp"
-#include "rclcpp/rclcpp.hpp"
-#include "rcpputils/filesystem_helper.hpp"
-#include "rcpputils/get_env.hpp"
-#include "sensor_msgs/msg/point_cloud2.hpp"
-#include "std_msgs/msg/float64.hpp"
-#include "std_msgs/msg/string.hpp"
-#include "rclcpp/generic_publisher.hpp"
-#include "rclcpp/generic_subscription.hpp"
-
-typedef std_msgs::msg::String message_type;
+#include <mqtt_client/srv/IsConnected.h>
+#include <rclcpp/rclcpp.hpp>
 
 
 /**
@@ -59,21 +48,20 @@ namespace mqtt_client {
  *
  * The MqttClient enables connected ROS-based devices or robots to
  * exchange ROS messages via an MQTT broker using the MQTT protocol.
- * This currently works only for a predefined ROS message.
- * You have to specify the ROS message type for ROS messages you wish to
+ * This works generically for any ROS message, i.e. there is no need
+ * to specify the ROS message type for ROS messages you wish to
  * exchange via the MQTT broker.
  */
 class MqttClient : public rclcpp::Node,
                    public virtual mqtt::callback,
                    public virtual mqtt::iaction_listener {
 
- public:
+ protected:
   /**
-   * @brief Constructor to initialize the node.
+   * @brief Initializes node.
    */
   MqttClient();
 
- protected:
   /**
    * @brief Loads ROS parameters from parameter server.
    */
@@ -143,17 +131,14 @@ class MqttClient : public rclcpp::Node,
    *
    * @param   path_string  (relative) path as string
    *
-   * @return  rcpputils::fs::path  path variable
+   * @return  std::filesystem::path  path variable
    */
-  rcpputils::fs::path resolvePath(const std::string& path_string);
-
+  std::filesystem::path resolvePath(const std::string& path_string);
 
   /**
    * @brief Initializes broker connection and subscriptions.
    */
   void setup();
-
-  void setupSubscriptions();
 
   /**
    * @brief Sets up the client connection options and initializes the client
@@ -167,40 +152,33 @@ class MqttClient : public rclcpp::Node,
   void connect();
 
   /**
-   * @brief Serializes and publishes a predefined ROS message to the MQTT
-   * broker.
+   * @brief Serializes and publishes a generic ROS message to the MQTT broker.
    *
-   * Before serializing the ROS message and publishing it to the MQTT broker,
-   * metadata on the ROS message type is extracted. This type information is
-   * also sent to the MQTT broker on a separate topic.
+   * TODO
    *
-   * The MQTT payload for the actual ROS message carries the following:
-   * - 0 or 1 (indicating if timestamp is injected (=1))
-   * - serialized timestamp (optional)
-   * - serialized ROS message
-   *
-   * @param   ros_msg    generic ROS message
-   * @param   ros_topic  ROS topic where the message was published
+   * @param   serialized_msg  generic serialized ROS message
+   * @param   ros_topic       ROS topic where the message was published
    */
-  void ros2mqtt(const std::shared_ptr<rclcpp::SerializedMessage> serialized_msg,
+  void ros2mqtt(const std::shared_ptr<rclcpp::SerializedMessage>& serialized_msg,
                 const std::string& ros_topic);
 
   /**
    * @brief Publishes a ROS message received via MQTT to ROS.
    *
-   * This utilizes the ShapeShifter stored for the MQTT topic on which the
-   * message was received. The ShapeShifter has to be configured to the ROS
-   * message type of the message. If the message carries an injected timestamp,
-   * the latency is computed and published.
-   *
-   * The MQTT payload is expected to carry the following:
-   * - 0 or 1 (indicating if timestamp is injected (=1))
-   * - serialized timestamp (optional)
-   * - serialized ROS message
+   * TODO
    *
    * @param   mqtt_msg       MQTT message
    */
   void mqtt2ros(mqtt::const_message_ptr mqtt_msg);
+
+  /**
+   * @brief Publishes a primitive message received via MQTT to ROS.
+   *
+   * TODO
+   *
+   * @param   mqtt_msg     MQTT message
+   */
+  void mqtt2primitive(mqtt::const_message_ptr mqtt_msg);
 
   /**
    * @brief Callback for when the client has successfully connected to the
@@ -238,8 +216,8 @@ class MqttClient : public rclcpp::Node,
    * @return true if client is connected to the broker
    * @return false if client is not connected to the broker
    */
-  bool isConnectedService(mqtt_client::srv::IsConnected::Request& request,
-                          mqtt_client::srv::IsConnected::Response& response);
+  bool isConnectedService(srv::IsConnected::Request& request,
+                          srv::IsConnected::Response& response);
 
   /**
    * @brief Callback for when the client receives a MQTT message from the
@@ -294,7 +272,7 @@ class MqttClient : public rclcpp::Node,
     std::string pass;  ///< password
     struct {
       bool enabled;  ///< whether to connect via SSL/TLS
-      rcpputils::fs::path
+      std::filesystem::path
         ca_certificate;  ///< public CA certificate trusted by client
     } tls;               ///< SSL/TLS-related variables
   };
@@ -305,10 +283,10 @@ class MqttClient : public rclcpp::Node,
   struct ClientConfig {
     std::string id;  ///< client unique ID
     struct {
-      bool enabled;                   ///< whether client buffer is enabled
-      int size;                       ///< client buffer size
-      rcpputils::fs::path directory;  ///< client buffer directory
-    } buffer;                         ///< client buffer-related variables
+      bool enabled;                     ///< whether client buffer is enabled
+      int size;                         ///< client buffer size
+      std::filesystem::path directory;  ///< client buffer directory
+    } buffer;                           ///< client buffer-related variables
     struct {
       std::string topic;         ///< last-will topic
       std::string message;       ///< last-will message
@@ -319,10 +297,10 @@ class MqttClient : public rclcpp::Node,
     double keep_alive_interval;  ///< keep-alive interval
     int max_inflight;            ///< maximum number of inflight messages
     struct {
-      rcpputils::fs::path certificate;  ///< client certificate
-      rcpputils::fs::path key;          ///< client private keyfile
-      std::string password;             ///< decryption password for private key
-    } tls;                              ///< SSL/TLS-related variables
+      std::filesystem::path certificate;  ///< client certificate
+      std::filesystem::path key;          ///< client private keyfile
+      std::string password;  ///< decryption password for private key
+    } tls;                   ///< SSL/TLS-related variables
   };
 
   /**
@@ -330,16 +308,16 @@ class MqttClient : public rclcpp::Node,
    */
   struct Ros2MqttInterface {
     struct {
-      rclcpp::GenericSubscription::SharedPtr
-        subscription;      ///< ROS subscriber
-      int queue_size = 1;  ///< ROS subscriber queue size
-      std::string msg_type;
-    } ros;                 ///< ROS-related variables
+      rclcpp::GenericSubscription::SharedPtr subscriber;  ///< generic ROS subscriber
+      std::string msg_type;                               ///< message type of subscriber
+      int queue_size = 1;                                 ///< ROS subscriber queue size
+    } ros;                                                ///< ROS-related variables
     struct {
       std::string topic;      ///< MQTT topic
       int qos = 0;            ///< MQTT QoS value
       bool retained = false;  ///< whether to retain MQTT message
     } mqtt;                   ///< MQTT-related variables
+    bool primitive = false;   ///< whether to publish as primitive message
     bool stamped = false;     ///< whether to inject timestamp in MQTT message
   };
 
@@ -351,14 +329,14 @@ class MqttClient : public rclcpp::Node,
       int qos = 0;  ///< MQTT QoS value
     } mqtt;         ///< MQTT-related variables
     struct {
-      std::string topic;  ///< ROS topic
-      rclcpp::GenericPublisher::SharedPtr
-        publisher;  ///< generic ROS publisher
-      rclcpp::Publisher<std_msgs::msg::Float64>::SharedPtr
-        latency_publisher;   ///< ROS publisher for latency
-      int queue_size = 1;    ///< ROS publisher queue size
+      std::string topic;                              ///< ROS topic
+      rclcpp::GenericPublisher::SharedPtr publisher;  ///< generic ROS publisher
+      rclcpp::Publisher latency_publisher;            ///< ROS publisher for latency
+      int queue_size = 1;                             ///< ROS publisher queue size
       bool latched = false;  ///< whether to latch ROS message
     } ros;                   ///< ROS-related variables
+    bool primitive = false;  ///< whether to publish as primitive message (if
+                             ///< coming from non-ROS MQTT client)
   };
 
  protected:
@@ -380,8 +358,7 @@ class MqttClient : public rclcpp::Node,
   /**
    * @brief ROS Service server for providing connection status
    */
-  rclcpp::Service<mqtt_client::srv::IsConnected>::SharedPtr
-    is_connected_service_;
+  rclcpp::Service<srv::IsConnected>::SharedPtr is_connected_service_;
 
   /**
    * @brief Status variable keeping track of connection status to broker
@@ -417,33 +394,28 @@ class MqttClient : public rclcpp::Node,
    * @brief MQTT2ROS connection variables sorted by MQTT topic
    */
   std::map<std::string, Mqtt2RosInterface> mqtt2ros_;
-
-  rclcpp::TimerBase::SharedPtr check_subscriptions_timer_;
 };
 
 
-template <typename ParameterT>
-bool MqttClient::loadParameter(const std::string& key, ParameterT& value) {
-  bool found = MqttClient::get_parameter(key, value);
+template <typename T>
+bool MqttClient::loadParameter(const std::string& key, T& value) {
+  bool found = private_node_handle_.getParam(key, value);
   if (found)
-    RCLCPP_DEBUG(rclcpp::get_logger("rclcpp"),
-                 "Retrieved parameter '%s' = '%s'", key.c_str(),
+    RCLCPP_DEBUG(rclcpp::get_logger(), "Retrieved parameter '%s' = '%s'", key.c_str(),
                  std::to_string(value).c_str());
   return found;
 }
 
 
-template <typename ParameterT>
-bool MqttClient::loadParameter(const std::string& key, ParameterT& value,
-                               const ParameterT& default_value) {
-  bool found = MqttClient::get_parameter_or(key, value, default_value);
+template <typename T>
+bool MqttClient::loadParameter(const std::string& key, T& value,
+                               const T& default_value) {
+  bool found = private_node_handle_.param<T>(key, value, default_value);
   if (!found)
-    RCLCPP_WARN(rclcpp::get_logger("rclcpp"),
-                "Parameter '%s' not set, defaulting to '%s'", key.c_str(),
-                std::to_string(default_value).c_str());
+    RCLCPP_WARN(rclcpp::get_logger(), "Parameter '%s' not set, defaulting to '%s'", key.c_str(),
+                 std::to_string(default_value).c_str());
   if (found)
-    RCLCPP_DEBUG(rclcpp::get_logger("rclcpp"),
-                 "Retrieved parameter '%s' = '%s'", key.c_str(),
+    RCLCPP_DEBUG(rclcpp::get_logger(), "Retrieved parameter '%s' = '%s'", key.c_str(),
                  std::to_string(value).c_str());
   return found;
 }
